% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aveytoolkit_ahtsummary.R
\name{ahtsummary}
\alias{ahtsummary}
\title{ahtsummary}
\usage{
ahtsummary(x, hlength = 5, tlength = 5, digits = 2, markdown = FALSE)
}
\arguments{
\item{x}{variable to summarize at head and tail}

\item{hlength}{number of rows to include at head}

\item{tlength}{number of rows to include at tail}

\item{digits}{digits to show for numeric values}
}
\value{
returns NULL invisibly
}
\description{
print out the head and tail as well as a data summary
}
\details{
knitr::kable is used if knitr is installed and `markdown == TRUE`
}
\author{
ESS R package, modified by Stefan Avey
}
