% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aveytoolkit_annotatePDFs.R
\name{annotatePDFs}
\alias{annotatePDFs}
\title{annotatePDFs}
\usage{
annotatePDFs(host = "localhost", inFiles, outFiles, titles = NULL,
  footers = NULL, combineMultiple = TRUE,
  removeMultiple = combineMultiple)
}
\arguments{
\item{host}{name of the host (e.g. <<user>>@<<IP Address>>)}

\item{inFiles}{a character vector giving the current PDF names}

\item{outFiles}{a character vector of filenames to save PDFs to. Only the first
value is used if \code{combineMultiple} is TRUE}

\item{titles}{a character vector of titles for each plot in the same order as
cytoscapeWindowList (default is to use the title slot in the
cytoscapeWindowList). Use NULL to omit a title. If only 1 title is given,
it will be put on all plots.}

\item{footers}{a character vector of footers for each plot in the same order
as cytoscapeWindowList (default is NULL for no footers). If only 1 footer
is given, it will be put on all plots.}

\item{combineMultiple}{Should multiple windows be combined into a single PDF with
one image per page? (Default TRUE)}

\item{removeMultiple}{Should the temporary files be removed (use only if combining PDFs)?}

\item{cytoscapeWindowList}{a list of cytoscapeWindow objects for the windows to
save}

\item{filenames}{a character vector of filenames to save PDFs to. Only the first
value is used if \code{combineMultiple} is TRUE}
}
\value{
an (invisible) list of exit codes for each bash operation
}
\description{
Add content on top of existing PDF files and can combine them into one file
using command line tools
}
\details{
This function is a wrapper around some shell commands that allows you to annotate
multiple PDF files and combine them into 1. This wrapper requires ghostscript
and coherent PDF (cpdf) \url{http://www.coherentpdf.com/} on the host machine
}
\examples{

## Create 2 dummy plots that I want to combine.
## NOTE: You would never do this because pdf() can create multiple pages
pdf("inFile1.pdf"); plot(1:10); dev.off()
pdf("inFile2.pdf"); plot(10:20); dev.off()

## Annotate some PDFs without combining
## BROKEN!
annotatePDFs(inFiles = paste0("inFile", 1:2, ".pdf"),
             outFiles = paste0("outFile", 1:2, ".pdf"),
             titles = paste0("testTitle", 1:2),
             footers = paste0("testFooter", 1:2),
             combineMultiple = FALSE,
             removeMultiple = FALSE)

## Still keeps the original files because removeMultiple is FALSE
annotatePDFs(inFiles = paste0("inFile", 1:2, ".pdf"),
             outFiles = "outFile.pdf",
             titles = paste0("testTitle", 1:2),
             footers = paste0("testFooter", 1:2),
             combineMultiple = TRUE,
             removeMultiple = FALSE)

## Removes the original input files
annotatePDFs(inFiles = paste0("inFile", 1:2, ".pdf"),
             outFiles = "outFile.pdf",
             titles = paste0("testTitle", 1:2),
             footers = paste0("testFooter", 1:2),
             combineMultiple = TRUE,
             removeMultiple = TRUE)
}
\author{
Stefan Avey
}

