% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aveytoolkit_AverageReplicates.R
\name{AverageReplicates}
\alias{AverageReplicates}
\title{AverageReplicates}
\usage{
AverageReplicates(eSubSet, numRep)
}
\arguments{
\item{eSubSet}{a matrix or data.frame of values with samples as columns}

\item{numRep}{the number of replicates}
}
\value{
a data.frame of averaged values with column names coming from the first of each of the replicates with .avg appended
}
\description{
This function averages replicates in a matrix or data.frame
}
\note{
Assumes that the replicates are all next to each other
}
\examples{
mat <- matrix(rnorm(1000), ncol=10) ## 10 columns of random uniform numbers
avgMat <- AverageReplicates(mat, numRep=2) ## average adjacent pairs of columns
}
\author{
Stefan Avey
}
\keyword{aveytoolkit}

