% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aveytoolkit_barplotCI.R
\name{barplotCI}
\alias{barplotCI}
\title{barplotCI}
\usage{
barplotCI(x, CIs = NULL, compareTo = 1, ...)
}
\arguments{
\item{x}{list of values to create a boxplot from}

\item{CIs}{the confidence intervals. Default is \code{NULL} and they will be calculated as the 95\% confidence interval.}

\item{compareTo}{non-negative integer specifying to which element of x should comparisons be made for significance. If 0, no significance will be added.}

\item{...}{other arguments passed to \code{barplot2} function}
}
\value{
Same as return from barplot2.
A numeric vector (or matrix, when \code{beside = TRUE}), say \code{mp}, giving the coordinates of _all_ the bar midpoints drawn, useful for adding to the graph.
If \code{beside} is true, use \code{colMeans(mp)} for the midpoints of each _group_ of bars, see example.
}
\description{
Create a barplot from the list x with one bar for each element of x
}
\author{
Christopher Bolen (creator) ; Stefan Avey (modified)
}

