% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aveytoolkit_browseIndex.R
\name{browseIndex}
\alias{browseIndex}
\title{browseIndex}
\usage{
browseIndex(package = NULL, lib.loc = NULL)
}
\arguments{
\item{package}{a string with the name of package to use. If no name is supplied, the most recently loaded package is used.}

\item{lib.loc}{a string of the directory name of the R library, or ‘NULL’.  The default value of ‘NULL’ corresponds to all libraries currently known.}
}
\value{
invisibly, the URL passed to browseURL (i.e. "file:" + <index_filename>)
}
\description{
Simple function to open HTML page of index in the default browser
}
\details{
Only works for a single package. Could improve to list an HTML page with multiple packages that you could then choose from.  Not currently implemented.  Borrows the conecpts from utils::browseVignettes
}
\examples{
browseIndex("utils")
}
\author{
Stefan Avey
}
\keyword{aveytoolkit}

