% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cbind.fill}
\alias{cbind.fill}
\title{resetPar}
\usage{
cbind.fill(...)
}
\arguments{
\item{...}{the objects, that will be converted to a list, to bind column-wise}
}
\value{
the cbind'ed objects passed in
}
\description{
Simple function to combine multiple objects by column while filling in NAs into extra rows created from differing lengths
}
\examples{
x<-matrix(1:10,5,2)
y<-matrix(1:16, 4,4)
z<-matrix(1:12, 2,6)

cbind.fill(x,y)
cbind.fill(x,y,z)
cbind.fill(mtcars, mtcars[1:10,])
}
\author{
Dimitris Rizopoulos and Tyler Rinker
}
\references{
\url{http://stackoverflow.com/questions/7962267/cbind-a-df-with-an-empty-df-cbind-fill}
}
\seealso{
\code{\link{cbind}}
}
\keyword{aveytoolkit}

