% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aveytoolkit_collapseDataset.R
\name{collapseDataset}
\alias{collapseDataset}
\title{collapseDataset}
\usage{
collapseDataset(exprsVals, platform = NULL, mapVector = NULL, oper = max,
  prefer = c("none", "up", "down"), singleProbeset = FALSE,
  returnProbes = FALSE, deProbes = NULL, debug = FALSE)
}
\arguments{
\item{exprsVals}{a matrix or data.frame of numeric values with rownames denoting the identifiers.}

\item{platform}{the microarray platform the data comes from for extracting the gene symbols}

\item{mapVector}{a named character vector with names specififying the current identifiers (probes matching the rownames of exprsVals) and the values of the vector specifying the gene symbols (or other identifier to collapse to).}

\item{oper}{the operation used to choose which probe when multiple probes map to the same gene.  Default is max which will calculate the maximum of the average.}

\item{prefer}{one of "none", "up", or "down", can be abbreviated.}

\item{singleProbeset}{If \code{TRUE}, the operation applies to the average over all conditions and all values for a gene will come from one probeset. Otherwise, if \code{FALSE}, the operation applies to the probesets over all conditions and the values for a gene may come from different probe sets . Default is \code{FALSE} for compatability reasons but \code{TRUE} is recommended.}

\item{returnProbes}{if \code{TRUE}, a list of the collapsed expression matrix and the probes are both returned (see return).}

\item{deProbes}{a list with named vectors "up" and "down" giving the names of up and downregulated probes}

\item{debug}{When TRUE, things will be printed out to help debug errors}
}
\value{
If returnProbes is \code{TRUE}, a list containing the collapsed dataset in $exprsVals and the probes chosen in $probeSets.  Otherwise, if returnProbes is \code{FALSE}, only the expression matrix is returned.
}
\description{
Collapses a dataset from probes to gene symbols.
}
\details{
This function is designed to work for microarray data but can work for any sort of numeric matrix for which multiple rows need to be collapsed.

If singleProbeset is set to \code{FALSE}, the default for compatability reasons but untested and not recommended, the values for each sample will be taken from the maximum across any probe that maps to that gene.  This means that a gene's expression values may be a composition of values from different probes rather than a single probe.   Most users will not need to use the `prefer` argument.  If prefer is "up", when multiple deProbes match the same gene, the upregulated will be chosen.  Similary for "down".  Default is "none" and the probe with the `oper` (default max) will be chosen.

Note that it is possible for multiple probes to have the same operation (\code{oper}) over all conditions and, in this case, I've decided arbitarily to choose the first one.
}
\examples{
## Trivial Example showing basic functionality
fakeExpr <- matrix(rnorm(50, mean=8, sd=1), ncol=5, nrow=10,
                   dimnames=list(probes=paste("probe", 1:10, sep='_'),
                     samples=paste("sample", LETTERS[1:5], sep='_')))
mv <- rep(paste("Gene", LETTERS[1:5], sep='_'), each=2) # mapVector
names(mv) <- rownames(fakeExpr)
res <- collapseDataset(fakeExpr, mapVector=mv, oper=max,
                       singleProbeset=TRUE, # recommend setting singleProbeset to TRUE
                       returnProbes=TRUE)
res$probes
## between probe_1 and probe_2, probe_2 was chosen for Gene_A
## between probe_3 and probe_4, probe_4 was chosen for Gene_B
## etc.

res$exprsVals                           # collapsed expression values

## only difference is in rownames, numbers are identical
all.equal(res$exprsVals, fakeExpr[res$probes,])
}
\author{
Christopher Bolen, Modified by Stefan Avey
}
\keyword{aveytoolkit}

