% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aveytoolkit_geomMean.R
\name{geomMean}
\alias{geomMean}
\title{geomMean}
\usage{
geomMean(x, na.rm = FALSE, zero.propagate = TRUE)
}
\arguments{
\item{x}{a vector of non-negative numeric values.}

\item{na.rm}{(optional) whether to remove \code{NA} values before calculation. Default is \code{FALSE}}

\item{zero.propagate}{(optional) logical specifying whether zeros should be progated such that if there are zeros in `x` 0 should be returned.}
}
\value{
the geometric mean of x
}
\description{
Calculate the geometric mean
}
\details{
This function handles negative values by returning `NaN` and zeros by returning `0` by default.
}
\examples{
x <- 1:10
x2 <- c(x, NA)
x3 <- -5:5
x4 <- 0:4

geomMean(x)
mean(x)

geomMean(x2)
mean(x2)

geomMean(x2, na.rm = TRUE)
mean(x2, na.rm = TRUE)

## NaN because because x3 contains negative values ##
geomMean(x3)
mean(x3)

geomMean(x4)
geomMean(x4, zero.propagate = FALSE)
mean(x4)
}
\references{
\url{http://stackoverflow.com/questions/2602583/geometric-mean-is-there-a-built-in}
}
\seealso{
\code{\link{exp}} \code{\link{sum}} \code{\link{log}}
}
\author{
Paul McMurdie, Stefan Avey
}
\keyword{aveytoolkit}
