% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aveytoolkit_get_bioc.R
\name{get_bioc}
\alias{get_bioc}
\title{get_bioc}
\usage{
get_bioc(...)
}
\arguments{
\item{...}{The unquoted names of Bioconductor packages to install}
}
\value{
a character vector containing the names of packages passed into the
    function (for which installation was attempted but not necessarily
    completed)
}
\description{
Get and install Bioconductor packages by name
}
\details{
This is a convenience function that prevents need to copy/paste
    from Bioconductor website to get the biocLite function
}
\examples{
\dontrun{
## Install Mfuzz and qpcrNorm Bioconductor packages
get_bioc(Mfuzz, qpcrNorm)
}
}
\author{
Robert Amezquita, Stefan Avey
}
