% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aveytoolkit_getBaseTheme.R
\name{getBaseTheme}
\alias{getBaseTheme}
\title{getBaseTheme}
\usage{
getBaseTheme()
}
\description{
\code{getBaseTheme} Defines universal plotting settings to use with ggplot2
}
\examples{
p <- ggplot(mtcars, aes(wt, mpg))
p <- p + geom_point() + getBaseTheme()
plot(p)
}
\author{
Jason Vander Heiden <jason.vanderheiden@yale.edu>
}
\keyword{aveytoolkit}
