% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aveytoolkit_GetObjectSizes.R
\name{GetObjectSizes}
\alias{GetObjectSizes}
\title{GetObjectSizes}
\usage{
GetObjectSizes(name = ".GlobalEnv", units = "Mb")
}
\arguments{
\item{name}{which environment to use in listing the available objects.
Defaults to the _current_ environment.  Although called
‘name’ for back compatibility, in fact this argument can
specify the environment in any form; see the ‘Details’
of \code{ls()} for more information.}

\item{units}{the units to be used in printing the size.  Allowed values
 are ‘"b"’, ‘"Kb"’, ‘"Mb"’, ‘"Gb"’, ‘"Tb"’, ‘"Pb"’, ‘"B"’,
‘"KB"’, ‘"MB"’, ‘"GB"’, ‘"TB"’, ‘"PB"’, ‘"KiB"’, ‘"MiB"’,
‘"GiB"’, ‘"TiB"’, ‘"PiB"’, ‘"EiB"’, ‘"ZiB"’, ‘"YiB"’, and
‘"auto"’ (see ‘Details’ of \code{object.size}).  Can be abbreviated.}
}
\value{
A named character vector with names corresponding to objects and values
        corresponding to strings in human-readable format
}
\description{
\code{GetObjectSizes} uses \code{ls()} and \code{object.size} to see what
objects are using most of the memory. \code{lsos()} is better for this purpose.
}
\details{

}
\examples{
## First Example
bigMat <- matrix(NA, nrow = 1000, ncol = 1000)
biggerMat <- matrix(NA, nrow = 10000, ncol = 10000)
GetObjectSizes()
GetObjectSizes(units = "Gb")
}
\author{
Stefan Avey
}
\seealso{
\code{ls}, \code{object.size}
}
\keyword{aveytoolkit}

