% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aveytoolkit_ggDualAxis.R
\name{ggDualAxis}
\alias{ggDualAxis}
\title{ggplot dual axis}
\usage{
ggDualAxis(plot1, plot2, which.axis = c("y", "x"))
}
\arguments{
\item{plot1}{a ggplot2 object}

\item{plot2}{a ggplot2 object}

\item{which.axis}{character vector of length 1 specifying "x" or "y" axis. Default if "y".}
}
\description{
\code{aveytoolkit_ggDualAxis} Takes two ggplot objects and combines them onto a single plot with dual x or y axes.
}
\examples{
## Example for x axis
# Create fake data.frame
data.add.x = data.frame(
  y1 = runif(100, 0, 100),
  x1 = runif(100, 0, 100)
)

# Add second x-axis that scales with first
data.add.x$x2 = (data.add.x$x1 + 50)^0.75

# Create plots
plot1.x = qplot(y = y1, x = x1, data = data.add.x)
plot2.x = qplot(y = y1, x = x2, data = data.add.x, col = 2)

# Run function
ggDualAxis(plot1.x, plot2.x, "x")

## Example for y axis
# Add second y-axis that scales with first
data.add.x$y2 = (data.add.x$y^0.5) / 500

# Create plots
plot1.y = qplot(y = y1, x = x1, data = data.add.x)
plot2.y = qplot(y = y2, x = x1, data = data.add.x)

# Run function
ggDualAxis(plot1.y, plot2.y, "y")
}
\author{
Jon Lefcheck
}
\references{
\url{https://gist.github.com/jslefche/e4c0e9f57f0af49fca87}
}

