% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aveytoolkit_ggSmartBoxplot.R
\name{ggSmartBoxplot}
\alias{ggSmartBoxplot}
\title{ggSmartBoxplot}
\usage{
ggSmartBoxplot(
  x,
  mat,
  splitRowBy = NA,
  splitColBy = NA,
  colorBy = NULL,
  rows,
  cols = NA,
  whichCols = NA,
  sep = ".",
  outlier.shape = 17,
  ylab = NULL,
  space = "fixed",
  scales = "fixed",
  fileName = NA,
  plot = TRUE,
  ...
)
}
\arguments{
\item{x}{the variable to group by for boxplots}

\item{mat}{data.frame or matrix of values to plot with samples in columns}

\item{splitRowBy}{a factor used to split the data by row in facet_grid}

\item{splitColBy}{a factor used to split the data by col in facet_grid}

\item{colorBy}{a factor used for coloring. No coloring will be done if \code{NULL} (default)}

\item{rows}{row names or row indices of the items to be plotted}

\item{cols}{substring to search for with "grep" in column names to be plotted}

\item{whichCols}{the column indices or full column names}

\item{sep}{a separator used in searching for cols in the column names}

\item{outlier.shape}{shape of outliers (default is 17, filled triangle)}

\item{ylab}{if NULL, default is to use rownames. Can specify a string instead to use}

\item{space}{If \code{"fixed"}, the default, all panels have the same size.  If \code{"free_y"} their height will be proportional to the length of the y scale; if \code{"free_x"} their width will be proportional to the length of the x scale; or if \code{"free"} both height and width will vary.  This setting has no effect unless the appropriate scales also vary.}

\item{scales}{Are scales shared across all facets (the default, \code{"fixed"}), or do they vary across rows (\code{"free_x"}), columns (\code{"free_y"}), or both rows and columns (\code{"free"})}

\item{fileName}{}

\item{plot}{logical specifying whether or not to plot the plot(s). Default is TRUE.}

\item{...}{other arguments that are passed to qplot}

\item{filename}{the name of a file to write a PDF to or \code{NA} to plot in standard graphics device.}
}
\value{
invisibly returns a list with 2 elements:
ggplot: the ggplot object to be plotted (this can be added to
dat: a named list of the data frame(s) passed to data in ggplot.  The names come from converting the rows argument to a character vector.
}
\description{
Boxplot wrapper for ggplot
}
\examples{
data(OrchardSprays)
## Example of functionality
ggSmartBoxplot(x=OrchardSprays$treatment,
               mat=t(OrchardSprays[,1]),
              rows=1, whichCols=1:ncol(t(OrchardSprays)),
              colorBy=factor(OrchardSprays$rowpos+OrchardSprays$colpos > 9),
              xlab="Treatment")

\dontrun{
cellType <- "PBMC"
## expr would be an expression matrix with genes in rows and samples in columns 
geneSub <- grep("HLA-A29.1", rownames(expr))
age <- "Young"
ages <- c("Young", "Old")
responses <- c("NR", "R")
subset <- targetFClist[[cellType]]$Age \%in\% ages & 
  targetFClist[[cellType]]$Response \%in\% responses
ggSmartBoxplot(x=targetFClist[[cellType]][subset, "Time"],
               mat=exprFClist[[cellType]], ylim=c(-1,1),
               rows=geneSub, whichCols=which(subset), 
               colorBy=targetFClist[[cellType]][subset,"Response"],
               splitRowBy=targetFClist[[cellType]][subset,"Age"],
               xlab="Days (Post Vaccination)",
               fileName=NA)
}
}
\seealso{
\code{\link{ggplot2}}, \code{\link{qplot}}
}
\author{
Stefan Avey
}
\keyword{aveytoolkit}
