% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aveytoolkit_ggSmoothExprPlot.R
\name{ggSmoothExprPlot}
\alias{ggSmoothExprPlot}
\title{ggSmoothExprPlot}
\usage{
ggSmoothExprPlot(x, mat, rows, method = "auto", formula = formula("y ~ x"),
  splitRowBy = NA, splitColBy = NA, colorBy = NULL, cols = NA,
  whichCols = NA, sep = ".", colorByLabel = "Response", ggtitle = TRUE,
  xlab = "Time (Days Post-Vaccination)", ylab = "Expression",
  colors = c("#1B9E77", "#D95F02", "#7570B3", "#E7298A", "#66A61E", "#E6AB02",
  "#A6761D", "#666666"), space = "fixed", scales = "fixed", fileName = NA,
  plot = TRUE)
}
\arguments{
\item{x}{the numeric x-axis variable for the plot (usually time)}

\item{mat}{data.frame or matrix of values to plot with samples in columns}

\item{rows}{row names or row indices of the items to be plotted}

\item{method}{smoothing method (function). See \code{stat_smooth}}

\item{formula}{a formula to use for smoothing in \code{stat_smooth} (e.g. the default "y ~ x" or "y ~ ns(x, 3)").}

\item{splitRowBy}{a factor used to split the data by row in facet_grid}

\item{splitColBy}{a factor used to split the data by col in facet_grid}

\item{colorBy}{a factor used for coloring. No coloring will be done if \code{NULL} (default)}

\item{cols}{substring to search for with "grep" in column names to be plotted}

\item{whichCols}{the column indices or full column names}

\item{sep}{a separator used in searching for cols in the column names}

\item{colorByLabel}{the labels used for the color legend}

\item{ggtitle}{logical. If TRUE, \code{rows} is coerced to character and passed to ggtitle}

\item{xlab}{passed to \code{xlab}. Defaults to "Time (Post-Vaccination)"}

\item{ylab}{passed to \code{ylab}. Defaults to "Expression"}

\item{colors}{The colors to use. Defaults to the colors given by using RColorBrewer's "Dark2" pallette (but RColorBrewer is not called directly so is not required).}

\item{space}{If \code{"fixed"}, the default, all panels have the same size.  If \code{"free_y"} their height will be proportional to the length of the y scale; if \code{"free_x"} their width will be proportional to the length of the x scale; or if \code{"free"} both height and width will vary.  This setting has no effect unless the appropriate scales also vary.}

\item{scales}{Are scales shared across all facets (the default, \code{"fixed"}), or do they vary across rows (\code{"free_x"}), columns (\code{"free_y"}), or both rows and columns (\code{"free"})}

\item{fileName}{the name of a file to write a PDF to or \code{NA} to plot in standard graphics device.}

\item{plot}{logical specifying whether or not to plot the plot(s). Default is TRUE.}
}
\value{
invisibly returns a list with 2 elements:
ggplot: the ggplot object to be plotted (this can be added to
dat: a named list of the data frame(s) passed to data in ggplot.  The names come from converting the rows argument to a character vector.
}
\description{
Wrapper around ggplot to transform data and plot profiles (e.g. expression or activity) over time
}
\examples{
data(OrchardSprays)
## Example of functionality
library(Biobase)
data(sample.ExpressionSet, package="Biobase")
dat <- sample.ExpressionSet
## Normally x-axis is time but in this dataset there is no time
## so we will use the `score` as the x-axis
genderF <- dat$sex == "Female"
ggSmoothExprPlot(x=dat$score[genderF],
                 mat=exprs(dat),
                 rows="31345_at",
                 whichCols=which(genderF), # females only
                 colorBy=as.factor(dat$type)[genderF],
                 colorByLabel="Condition",
                 xlab="score")
\dontrun{tmp <- ggSmoothExprPlot(x=times[subset], mat=expr, rows=gene,
                                 formula=formula("y ~ ns(x,3)"),
                                 whichCols=subset, colorBy=target[subset,respType],
                                 splitColBy=splitby,
                                 splitRowBy=as.factor(target[subset,"Study"]),
                                 ggtitle=TRUE, colorByLabel=respType, plot=TRUE)}
}
\author{
Stefan Avey
}
\seealso{
\code{\link{ggplot2}}
}
\keyword{aveytoolkit}

