% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aveytoolkit_INT.R
\name{INT}
\alias{INT}
\title{INT}
\usage{
INT(x, na.last = "keep", ties.method = c("average", "first", "last",
  "random", "max", "min"), ...)
}
\arguments{
\item{x}{numeric vector to be transformed}

\item{na.last}{How NA values should be handled. Passed to \code{rank}.}

\item{ties.method}{How ties should be handled. Passed to \code{rank}.}

\item{...}{Other arguments passed to \code{qnorm}}
}
\value{
A numeric vector containing the transformed values of x.
}
\description{
\code{INT} INT performs an inverse normal transformation
}
\details{
Takes an input vector and performs a rank-based inverse normal transformation (making the data approximately normally distributed. Positions with missing (NA) values will be returned as NA by default (see `na.last`)
}
\examples{
## Normally Distributed data
x1 <- rnorm(100)
hist(INT(x1)) # still normally distributed
hist(INT(x1, mean = 10, sd = 2)) # still normally distributed

## Uniformly Distributed data
x2 <- runif(100)
hist(INT(x2)) # forced to be normally distributed by rank

## Many ties in data, different methods for handling ties
x3 <- rep(10:20, 5)
hist(INT(x3, ties.method = "average"))
hist(INT(x3, ties.method = "first"))
hist(INT(x3, ties.method = "max"))
}
\author{
Stefan Avey
}
\keyword{aveytoolkit}
