% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aveytoolkit_lsos.R
\name{lsos}
\alias{lsos}
\title{lsos}
\usage{
lsos(..., n = 10)
}
\description{
\code{lsos} will list objects by size to help manage memory in R
}
\examples{
## First Example
bigMat <- matrix(NA, nrow = 1000, ncol = 1000)
biggerMat <- matrix(NA, nrow = 10000, ncol = 10000)
lsos()
}
\seealso{
\code{ls}, \code{object.size}
}
\author{
Dirk Eddelbuttel
}
