% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aveytoolkit_MakeDF.R
\name{MakeDF}
\alias{MakeDF}
\title{MakeDF}
\usage{
MakeDF(List, Names)
}
\arguments{
\item{List}{the list to convert}

\item{Names}{the names of the list}
}
\value{
a data frame of the converted list.
}
\description{
Creates a data frame from a list.  Useful for when the list elements have unequal lengths and
\code{\link{as.data.frame}} fails.
}
\references{
\url{http://stackoverflow.com/questions/15753091/convert-mixed-length-named-list-to-data-frame}
}
\seealso{
\code{\link{gsub}}
}
\author{
Josh O'Brien
}
\keyword{aveytoolkit}
