% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aveytoolkit_makeTransparent.R
\name{makeTransparent}
\alias{makeTransparent}
\title{makeTransparent}
\usage{
makeTransparent(alpha = 0.5, ...)
}
\arguments{
\item{alpha}{transparency factor in range [0,1]}

\item{...}{vector or list of colors}
}
\value{
a vector of new colors made transparent
}
\description{
Simple function to make some colors transparent
}
\examples{
makeTransparent("red", "blue")
##[1] "#FF00007F" "#0000FF7F"
makeTransparent("red", "blue", alpha=0.8)
## [1] "#FF0000CC" "#0000FFCC"
}
\author{
Ricardo Oliveros-Ramos
}
\references{
\url{http://stackoverflow.com/questions/8047668/transparent-equivalent-of-given-color}
}
\seealso{
\code{\link{rgb}}, \code{\link{col2rgb}}
}
\keyword{aveytoolkit}

