% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aveytoolkit_Pause.R
\name{Pause}
\alias{Pause}
\title{Pause}
\usage{
Pause(str = "continue", quiet = FALSE)
}
\arguments{
\item{str}{optional string to display. Defaults to "continue".}

\item{quiet}{if TRUE, no prompt is displayed. Default is FALSE}
}
\value{
NULL is returned by invisible
}
\description{
This function prompts for return key and waits until the return is pushed to continue execution. It is used often to view plots coded in a loop one at a time allowing the user to control when the next plot should be displayed
}
\details{
The Pause function uses readline to wait until a newline character (produced by the Enter key) is given.  Instead of pressing Enter, a newline character can be used to automate this waiting time.
}
\examples{
for(p in 1:10) {
  plot(-10:10, (-10:10)^p, type='b')
  Pause(paste0('see plot of x^',p+1))
}
}
\seealso{
\code{\link{readline}}, \code{\link{invisible}}
}
\author{
Stefan Avey
}
\keyword{aveytoolkit}
