% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aveytoolkit_ProcessNames.R
\name{ProcessNames}
\alias{ProcessNames}
\title{ProcessNames}
\usage{
ProcessNames(strs, stringsToRm = NULL, rmPunct = TRUE, sep = "_")
}
\arguments{
\item{strs}{vector or strings to process}

\item{stringsToRm}{a vector or list of strings to search for and remove from strs}

\item{rmPunct}{should punctuation be removed? Default is TRUE.}

\item{sep}{character to replace whitespace}
}
\value{
a vector of modified strings from strs
}
\description{
Cleans up strings to make them pretty names by removing punctuation, whitespace, and specified substrings
}
\details{
stringsToRm are replaced by '' in the order they are given using gsub. After this, punctuation is removed if rmPunct is TRUE. Then, leading and/or trailing whitespace will be removed and the sep will be used to separate words. This function is useful when reading in other people's data and you want to change the row or column names to legal R names or just shorten the names.
}
\examples{
badNames <- c("Who's Birthday?", "[Date]", "gift Received")
## Remove the string "Who's", remove punctuation, and separate words by '_'
goodNames <- ProcessNames(badNames,stringsToRm="Who's", rmPunct=TRUE, sep='_')
goodNames
## Remove the string "Who's", don't remove punctuation, and put no separation between words
goodNames <- ProcessNames(badNames,stringsToRm="Who's", rmPunct=FALSE, sep='')
goodNames
}
\author{
Stefan Avey
}
\seealso{
\code{\link{gsub}}
}
\keyword{aveytoolkit}

