% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aveytoolkit_readGMT.R
\name{readGMT}
\alias{readGMT}
\title{readGMT.R}
\usage{
readGMT(filename, trimMissing = TRUE, quiet = FALSE)
}
\arguments{
\item{filename}{the GTM file to read in (must end in .gmt)}

\item{trimMissing}{logical indicating whether to trim missing gene identifiers that are read as empty strings (default is TRUE)}

\item{quiet}{logical indicating whether to show how many records read or stay quiet. Default is to pass quiet = FALSE to `scan()`}
}
\value{
a list with the following elements

\describe{
  \item{genesets}{a list with one element per gene set containing a character vector of genes}
  \item{names}{a list with one element per gene set containing the set names}
  \item{descriptions}{a list with one element per gene set containing the set descriptions}
}
}
\description{
Read in a GMT (Gene Matrix Transposed) file.
}
\details{
Read in a vector of set names, descriptions, and gene identifiers and
         store them in a list. For more details on the GTM format, see
         \url{http://www.broadinstitute.org/cancer/software/gsea/wiki/index.php/Data_formats#GMT:_Gene_Matrix_Transposed_file_format_.28.2A.gmt.29}. This code is heavily adapted from `qusage::read.gmt()` and `GSA::GSA.read.gmt()` to take the best of both worlds.
}
\author{
Stefan Avey.
}

