% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aveytoolkit_resetPar.R
\name{resetPar}
\alias{resetPar}
\title{resetPar}
\usage{
resetPar()
}
\value{
an invisible named list of parameters returned by calling par
}
\description{
Simple function to reset plotting parameters for when things get wonky
}
\details{
This function resets the graphical parameters from the par function. It flashes a new device on the screen but works to reset parameters.  Meant to be used when things get hairy and not coded in scripts
}
\examples{
par(oma=c(4,10,2,1))
plot(1,1)
## paramter settings weren't saved so do a reset
resetPar()
plot(1,1)
}
\author{
Gavin Simpson
}
\references{
\url{http://stackoverflow.com/questions/5789982/reset-par-to-the-default-values-at-startup}
}
\seealso{
\code{par}
}
\keyword{aveytoolkit}

