% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aveytoolkit_VennDiagram.R
\name{VennDiagram}
\alias{VennDiagram}
\title{VennDiagram}
\usage{
VennDiagram(setList, mar = c(0, 0, 1, 0), ...)
}
\arguments{
\item{setList}{a (named) list of the sets to be plotted.  The names will be used on the plot. If the list is unnamed, the default names in \code{\link{vennDiagram}}}
}
\value{
a data frame of binary values indicating membership in each set with rownames giving the set entries.
}
\description{
Draw a venn diagram of 2 or 3 sets
}
\details{
Wrapper around the \code{\link{limma}} \code{\link{vennDiagram}} function to make it simpler.
}
\author{
Stefan Avey
}
\references{
Code modified from \url{http://research.stowers-institute.org/mcm/venn.R}
}
\seealso{
\code{\link{vennDiagram}}
}
\keyword{aveytoolkit}

