% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aveytoolkit_writeGMT.R
\name{writeGMT}
\alias{writeGMT}
\title{aveytoolkit_writeGMT.R}
\usage{
writeGMT(filename, sets, setNames = names(sets), setDescriptions = rep(NA,
  length(sets)))
}
\arguments{
\item{filename}{the file to write to (should include '.gmt' extension}

\item{sets}{a list of character vectors containing the sets to write}

\item{setNames}{a character vector of set names corresponding to sets.
Defaults to 1, 2, 3, ..., length(sets) if nonte specified.}

\item{setDescriptions}{a character vector of set descriptions corresponding
to sets. Defaults to \code{NA} values if none specified.}
}
\description{
Write out a GMT (Gene Matrix Transposed) file.
}
\details{
Take in a vector of set names, descriptions, and gene identifiers and
         write them to a GMT file format.
         http://www.broadinstitute.org/cancer/software/gsea/wiki/index.php/Data_formats#GMT:_Gene_Matrix_Transposed_file_format_.28.2A.gmt.29
}
\author{
Stefan Avey
}
