% Generated by roxygen2 (4.0.1): do not edit by hand
\name{collapseDataset}
\alias{collapseDataset}
\title{collapseDataset}
\usage{
collapseDataset(exprsVals, platform = NULL, mapVector = NULL, oper = max,
  prefer = c("none", "up", "down"), singleProbeset = FALSE,
  returnProbes = FALSE, deProbes = NULL)
}
\arguments{
\item{exprsVals}{a matrix or data.frame of expression values with rownames denoting the probes.}

\item{platform}{the microarray platform the data comes from for extracting the gene symbols}

\item{mapVector}{a named character vector with names specififying the current identifiers (probes matching the rownames of exprsVals) and the values of the vector specifying the gene symbols (or other identifier to collapse to).}

\item{oper}{the operation used to choose which probe when multiple probes map to the same gene.  Default is max which will calculate the maximum of the average.}

\item{prefer}{one of "none", "up", or "down", can be abbreviated.}

\item{singleProbeset}{If \code{TRUE}, the operation applies to the average of each sample. Otherwise, if \code{FALSE}, the operation applies to the probesets over all samples and only one probeset will be selected. Default is \code{FALSE}.}

\item{returnProbes}{if \code{TRUE}, a list of the collapsed expression matrix and the probes are both returned (see return).}

\item{deProbes}{a list with named vectors "up" and "down" giving the names of up and downregulated probes}
}
\value{
If returnProbes is \code{TRUE}, a list containing the collapsed dataset in $exprsVals and the probes chosen in $probeSets.  Otherwise, if returnProbes is \code{FALSE}, only the expression matrix is returned.
}
\description{
Collapses a dataset from probes to gene symbols.
}
\details{
If singleProbeset is set to \code{TRUE}, untested and not recommended, the values for each sample will be taken from the maximum across any probe that maps to that gene.  This means that a gene's expression values may be a composition of values from different probes rather than a single probe.   if prefer is "up", when multiple deProbes match the same gene, the upregulated will be chosen.  Similary for "down".  Default is "none" and the probe with the oper will be chosen.
}
\examples{
## ??
}
\author{
Christopher Bolen, Modified by Stefan Avey
}
\keyword{aveytoolkit}

