% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aveytoolkit_geomMean.R
\name{geomMean}
\alias{geomMean}
\title{resetPar}
\usage{
geomMean(x, na.rm = TRUE)
}
\arguments{
\item{x}{a vector of numeric values}

\item{na.rm}{(optional) whether to remove \code{NA} values before calculation. Default is \code{TRUE}}
}
\value{
the geometric mean of x
}
\description{
Calculate the geometric mean
}
\details{
This function handles 0 values by ignoring them
}
\examples{
x <- 1:10
x2 <- x^2
geomMean(x)
mean(x)
geomMean(x2)
mean(x2)
}
\author{
Paul McMurdie
}
\references{
\url{http://stackoverflow.com/questions/2602583/geometric-mean-is-there-a-built-in}
}
\seealso{
\code{\link{exp}} \code{\link{sum}} \code{\link{log}}
}
\keyword{aveytoolkit}

