% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aveytoolkit_getLoginDetails.R
\name{getLoginDetails}
\alias{getLoginDetails}
\title{getLoginDetails}
\usage{
getLoginDetails()
}
\value{
an invisible named vector of loginID and password
}
\description{
Uses tcltk to display a prompt for a loginID and password
}
\details{
This function displays a window for a user to enter a loginID and password without showing the password.
}
\examples{
credentials <- getLoginDetails()
## Do what needs to be done with loginID and password
rm(credentials) # Delete credentials
}
\author{
Markus Gesmann, Barry Rowlingson
}
\references{
\url{http://www.r-bloggers.com/simple-user-interface-in-r-to-get-login-details/} \url{http://r.789695.n4.nabble.com/tkentry-that-exits-after-RETURN-tt854721.html#none}
}
\seealso{
\code{\link{tcltk}}
}
\keyword{aveytoolkit}

