% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aveytoolkit_ggSmartBoxplot.R
\name{ggSmartBoxplot}
\alias{ggSmartBoxplot}
\title{ggSmartBoxplot}
\usage{
ggSmartBoxplot(x, mat, splitRowBy = NA, splitColBy = NA, colorBy = NULL,
  rows, cols = NA, whichCols = NA, sep = ".", outlier.shape = 17,
  outlier.color = NULL, fileName = NA, plot = TRUE, ...)
}
\arguments{
\item{x}{the variable to group by for boxplots}

\item{mat}{data.frame or matrix of values to plot with samples in columns}

\item{splitRowBy}{a factor used to split the data by row in facet_grid}

\item{splitColBy}{a factor used to split the data by col in facet_grid}

\item{colorBy}{a factor used for coloring. No coloring will be done if \code{NULL} (default)}

\item{rows}{row names or row indices of the items to be plotted}

\item{cols}{substring to search for with "grep" in column names to be plotted}

\item{whichCols}{the column indices or full column names}

\item{sep}{a separator used in searching for cols in the column names}

\item{outlier.shape}{shape of outliers (default is 17, filled triangle)}

\item{outlier.color}{color of outliers (default is NULL)}

\item{fileName}{}

\item{plot}{logical specifying whether or not to plot the plot(s). Default is TRUE.}

\item{...}{other arguments that are passed to qplot}

\item{filename}{the name of a file to write a PDF to or \code{NA} to plot in standard graphics device.}
}
\value{
invisibly returns a named list of the data frame(s) used for plotting the boxplot(s).  The names come from converting the rows argument to a character vector.
}
\description{
Boxplot wrapper for ggplot
}
\examples{
data(OrchardSprays)
## Example of functionality
ggSmartBoxplot(x=OrchardSprays$treatment,
               mat=t(OrchardSprays[,1]),
              rows=1, whichCols=1:ncol(t(OrchardSprays)),
              colorBy=factor(OrchardSprays$rowpos+OrchardSprays$colpos > 9),
              xlab="Treatment")

## NOT RUN:
cellType <- "PBMC"
geneSub <- grep("HLA-A29.1", rownames(expr))
age <- "Young"
ages <- c("Young", "Old")
responses <- c("NR", "R")
subset <- targetFClist[[cellType]]$Age \%in\% ages &
  targetFClist[[cellType]]$Response \%in\% responses
ggSmartBoxplot(x=targetFClist[[cellType]][subset, "Time"],
               mat=exprFClist[[cellType]], ylim=c(-1,1),
               rows=geneSub, whichCols=which(subset),
               colorBy=targetFClist[[cellType]][subset,"Response"],
               splitRowBy=targetFClist[[cellType]][subset,"Age"],
               xlab="Days (Post Vaccination)",
               fileName=NA)
}
\author{
Stefan Avey
}
\seealso{
\code{\link{ggplot2}}, \code{\link{qplot}}
}
\keyword{aveytoolkit}

