% Generated by roxygen2 (4.0.1): do not edit by hand
\name{runLimma}
\alias{runLimma}
\title{runLimma}
\usage{
runLimma(eset, labels, contrasts, block = NULL, covariates = NULL,
  filterReplicateGenes = TRUE, min.fold.change = 1, min.intensity = 4,
  p.cutoff = 0.05, fitOnly = FALSE)
}
\arguments{
\item{eset}{the expression matrix}

\item{labels}{the labels for each column of the eset}

\item{contrasts}{Vector of contrasts to make}

\item{block}{vector or factor specifying a blocking variable on the arrays. Has length equal to the number of arrays. Must be ‘NULL’ if ‘ndups>2’. (Not extensively tested, use with caution)}

\item{covariates}{data frame of covariates (of same length as labels) to include in the model. Use this if there are paired samples, etc.}

\item{filterReplicateGenes}{Only include one probeset for each gene (determined by symbol)}

\item{min.fold.change}{Minimum log2 fold change to be differentially expressed. Default is 1.}

\item{min.intensity}{Minimum log2 intensity (at any time) to be differentially expressed. Default is 4.}

\item{p.cutoff}{FDR corrected cutoff for significant differential expression. Default is 0.05.}

\item{fitOnly}{If true, will return fit2, rather than the matrix of significant genes. Default is FALSE.}
}
\value{
depends on \code{fitOnly}
}
\description{
A wrapper around limma functions to perform a basic analysis on the given expression matrix
}
\details{
Generally, an expression matrix is made up of rows of genes (or any other features) and columns of samples. The matrix has data for multiple classes (which are denoted with the 'labels' parameter) and the classes are compared using the vector of contrasts. Block can be used for biological (or technical) replicates or for separate subjects (in which case it will determinte the inter-subject correlation).  See \code{?duplicateCorrelation} for more information.
}
\examples{
## Example:
## If you have a m X 10 matrix 'eset', with 5 samples of class A and 5 of class B,
## you could compare class A to class B using the following code:
##
## results = runLimma(eset, c('A','A','A','A','A','B','B','B','B','B'), "B-A")
##
## This will return to you a matrix with columns for each comparison and rows for each gene.
## The value in each cells will either be -1, 0, or 1, depending on whether the gene is
## significantly higher in B, not significant, or significantly higher in A, respectively.
## If you want information on p-values and fold changes, set "fitOnly=T", and you can access
## the fit object to get the information.
##
## For other comparisons, you can look at the LIMMA user guide: limmaUsersGuide()
}
\author{
Christopher Bolen, Stefan Avey
}
\seealso{
\code{\link{limma}}
}
\keyword{aveytoolkit}

